# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Internet Bank Backend API
  version: 1.1.0
  description: |
    Актуальная HTTP-спецификация backend-сервисов интернет-банка через API Gateway.
    Все запросы отправляются на один базовый URL.
servers:
  - url: http://localhost:8080
    description: API Gateway
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Account:
      type: object
      required: [id, client_id, balance, currency, status, opened_at]
      properties:
        id:
          type: string
          format: uuid
        client_id:
          type: string
          format: uuid
        balance:
          type: number
          format: double
        currency:
          type: string
          enum: [RUB, USD, EUR]
        status:
          type: string
          enum: [active, closed]
        opened_at:
          type: string
          format: date-time
        closed_at:
          type: string
          format: date-time
          nullable: true
    Operation:
      type: object
      required: [id, account_id, type, amount, balance_after, created_at]
      properties:
        id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [deposit, withdraw, credit_issue, credit_repay]
        amount:
          type: number
          format: double
        balance_after:
          type: number
          format: double
        created_at:
          type: string
          format: date-time
        description:
          type: string
        credit_id:
          type: string
          format: uuid
          nullable: true
...
    AccountListResponse:
      type: object
      required: [accounts, pageNumber, pageQuantity]
      properties:
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/Account'
        pageNumber:
          type: integer
        pageQuantity:
          type: integer
    OperationListResponse:
      type: object
      required: [operations, pageNumber, pageQuantity]
      properties:
        operations:
          type: array
          items:
            $ref: '#/components/schemas/Operation'
        pageNumber:
          type: integer
        pageQuantity:
          type: integer
...
    TransferResponse:
      type: object
      required: [debit_operation, credit_operation]
      properties:
        debit_operation:
          $ref: '#/components/schemas/Operation'
        credit_operation:
          $ref: '#/components/schemas/Operation'
    OpenAccountRequest:
      type: object
      required: [client_id]
      properties:
        client_id:
          type: string
          format: uuid
        currency:
          type: string
          enum: [RUB, USD, EUR]
          default: RUB
    ChangeBalanceRequest:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
          format: double
          minimum: 0.01
    TransferRequest:
      type: object
      required: [from_account_id, to_account_id, amount]
      properties:
        from_account_id:
          type: string
          format: uuid
        to_account_id:
          type: string
          format: uuid
        amount:
          type: number
          format: double
          minimum: 0.01
...
paths:
...
  /accounts:
    post:
      tags: [Accounts]
      summary: Открыть счет
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenAccountRequest'
      responses:
        '201':
          description: Счет создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
    get:
      tags: [Accounts]
      summary: Список счетов
      parameters:
        - in: query
          name: client_id
          schema:
            type: string
            format: uuid
        - in: query
          name: status
          schema:
            type: string
            enum: [active, closed]
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 50
            minimum: 1
      responses:
        '200':
          description: Список счетов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountListResponse'
  /accounts/{accountId}:
    get:
      tags: [Accounts]
      summary: Получить счет
      parameters:
        - in: path
          name: accountId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Счет
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
    delete:
      tags: [Accounts]
      summary: Закрыть счет
      parameters:
        - in: path
          name: accountId
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: client_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Счет закрыт
  /accounts/{accountId}/deposit:
    post:
      tags: [Operations]
      summary: Пополнение счета
      parameters:
        - in: path
          name: accountId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeBalanceRequest'
      responses:
        '200':
          description: Создана операция
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
  /accounts/{accountId}/withdraw:
    post:
      tags: [Operations]
      summary: Списание со счета
      parameters:
        - in: path
          name: accountId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeBalanceRequest'
      responses:
        '200':
          description: Создана операция
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Operation'
  /accounts/transfer:
    post:
      tags: [Operations]
      summary: Перевод между счетами (с FX-конвертацией при разных валютах)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '200':
          description: Возвращает дебетовую и кредитовую операции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResponse'
  /accounts/{accountId}/operations:
    get:
      tags: [Operations]
      summary: История операций по счету
      parameters:
        - in: path
          name: accountId
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 50
            minimum: 1
      responses:
        '200':
          description: Список операций
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationListResponse'
  /ws/accounts/{accountId}/operations:
    get:
      tags: [Operations]
      summary: WebSocket stream операций по счету
      description: |
        Требует `Authorization: Bearer <token>` при WS-handshake.
        Сначала сервер отправляет `operations_snapshot`, затем события `operation_created` в реальном времени.
      parameters:
        - in: path
          name: accountId
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
        - in: query
          name: page_size
          schema:
            type: integer
            default: 50
            minimum: 1
      responses:
        '101':
          description: Protocol Switch to WebSocket
...